parser/requirements.txt
psycopg2-binary
requests
pandas
numpy
netcdf4
xarray
watchdog>=3.0.0
python-dateutil>=2.8.0
pytest
pytest-cov